move quest_watcher into a separate module
