+++ lib/quest.rb
@@ -4,6 +4,7 @@ require 'json'
 require 'grape'
 require 'quest/logger'
 require 'quest/messenger'
+require 'quest/spec_runner'
 require 'quest/quest_watcher'
 require 'quest/colorization'
 require 'quest/api'
